Towards zero default and hundred purcent coverage

+++ src/fr/ut1c/m2ipm/portfolio/Portfolio.java
@@ -97,7 +97,7 @@ public final class Portfolio {
     public float getValue(final Day day) {
         double total = 0;
         for (PortfolioLine lp : this.mapLines.values()) {
-            total = total + (lp.getQuantity() * lp.getShare().getPrice(day));
+            total = total + ((double) lp.getQuantity() * (double) lp.getShare().getPrice(day));
         }
         return (float) total;
     }
